Gen_verif_hash/requirements.txt
bcrypt>=4.2.0
argon2-cffi>=23.1.0

